Changed event arguments type from int to size_t

+++ neo/d3xp/gamesys/Class.cpp
@@ -828,7 +828,7 @@ idClass::ProcessEventArgs
 bool idClass::ProcessEventArgs( const idEventDef *ev, int numargs, ... ) {
 	idTypeInfo	*c;
 	int			num;
-	int			data[ D_EVENT_MAXARGS ];
+	size_t			data[ D_EVENT_MAXARGS ];
 	va_list		args;
 	
 	assert( ev );
@@ -936,7 +936,7 @@ bool idClass::ProcessEvent( const idEventDef *ev, idEventArg arg1, idEventArg ar
 idClass::ProcessEventArgPtr
 ================
 */
-bool idClass::ProcessEventArgPtr( const idEventDef *ev, int *data ) {
+bool idClass::ProcessEventArgPtr( const idEventDef *ev, size_t *data ) {
 	idTypeInfo	*c;
 	int			num;
 	eventCallback_t	callback;
